refacto the feature by using a dedicated component
