+++ frontend/src/components/Navbar.tsx
@@ -1,29 +1,9 @@
 import { Link, useLocation } from 'react-router-dom';
 import { Button } from '@/components/ui';
-import {
-  TrophyIcon,
-  Leaf,
-  HomeIcon,
-  UserRound,
-  BarChart3,
-  Sun,
-  Moon,
-  Laptop,
-  LogOutIcon,
-} from 'lucide-react';
+import { TrophyIcon, Leaf, HomeIcon, BarChart3 } from 'lucide-react';
 import { cn } from '@/lib/utils';
 
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from '@/components/ui/dropdown-menu';
-
-import { useUserStore } from '@/lib/zustand/userStore';
-import { useTheme } from './theme-provider';
+import { AccountDropdown } from './AccountDropdown';
 
 const navItems = [
   { to: '/challenges', icon: TrophyIcon, label: 'Challenges' },
@@ -34,17 +14,9 @@ const navItems = [
 
 export const Navbar = () => {
   const location = useLocation();
-  const isAuth = useUserStore((state) => !!state.user);
-  const { theme, setTheme } = useTheme();
-
-  const themeIcon = {
-    light: <Sun className="h-4 w-4" />,
-    dark: <Moon className="h-4 w-4" />,
-    system: <Laptop className="h-4 w-4" />,
-  };
 
   return (
-    <nav className="h-navbar bg-sidebar text-sidebar-foreground fixed bottom-0 z-50 flex w-full items-center justify-around rounded-t-2xl px-2 sm:hidden">
+    <nav className="h-navbar bg-sidebar text-sidebar-foreground pointer-events-auto fixed bottom-0 z-50 flex w-full items-center justify-around rounded-t-2xl px-2 sm:hidden">
       {navItems.map(({ to, icon: Icon, label }) => {
         const isActive = location.pathname === to;
 
@@ -66,62 +38,8 @@ export const Navbar = () => {
           </Button>
         );
       })}
-      <DropdownMenu>
-        <DropdownMenuTrigger>
-          <UserRound className="bg-sidebar-primary text-sidebar-primary-foreground hover:bg-sidebar-accent hover:text-sidebar-accent-foreground size-12 cursor-pointer rounded-md p-3 transition-colors" />
-        </DropdownMenuTrigger>
-        <DropdownMenuContent className="w-56">
-          {isAuth ? (
-            <>
-              <DropdownMenuLabel>Mon profil</DropdownMenuLabel>
-              <DropdownMenuSeparator />
-              <DropdownMenuItem asChild>
-                <Link to="/user" className="w-full">
-                  Accéder à mon espace <UserRound className="size-4" />
-                </Link>
-              </DropdownMenuItem>
-              <DropdownMenuItem asChild>
-                <Link to="/user" className="flex justify-between gap-2">
-                  Déconnexion <LogOutIcon className="size-4" />
-                </Link>
-              </DropdownMenuItem>
-            </>
-          ) : (
-            <>
-              <DropdownMenuItem asChild>
-                <Link to="/user" className="w-full">
-                  Inscription / Connexion
-                </Link>
-              </DropdownMenuItem>
-            </>
-          )}
-          <DropdownMenuSeparator />
-          <DropdownMenuLabel className="flex items-center gap-2">
-            Thème: {themeIcon[theme]}
-          </DropdownMenuLabel>
-          <DropdownMenuSeparator />
-          <DropdownMenuItem
-            onClick={(e) => {
-              setTheme('light');
-              e.preventDefault();
-            }}
-          >
-            Thème clair {themeIcon['light']}
-          </DropdownMenuItem>
-          <DropdownMenuItem
-            onClick={(e) => {
-              setTheme('dark');
-              e.preventDefault();
-            }}
-          >
-            Thème sombre {themeIcon['dark']}
-          </DropdownMenuItem>
 
-          <DropdownMenuItem onClick={() => setTheme('system')}>
-            Préférence système {themeIcon['system']}
-          </DropdownMenuItem>
-        </DropdownMenuContent>
-      </DropdownMenu>
+      <AccountDropdown />
     </nav>
   );
 };
